introduce cartesian.plot method:
- handles plot across cartesian subplots, in a similar way
  to how gl3d, geo and gl2d module operate.
